move function from utils to state_inference
