+++ label_studio/core/settings/base.py
@@ -933,7 +933,7 @@ DM_MAX_USERS_TO_DISPLAY = int(get_env('DM_MAX_USERS_TO_DISPLAY', 10))
 # Base FSM (Finite State Machine) Configuration for Label Studio
 FSM_CACHE_TTL = 300  # Cache TTL in seconds (5 minutes)
 FSM_SYNC_PROJECT_STATE = 'fsm.project_transitions.sync_project_state'
-FSM_INFERENCE_FUNCTION = 'fsm.utils._get_or_infer_state'
+FSM_INFERENCE_FUNCTION = 'fsm.state_inference._get_or_infer_state'
 
 # Used for async migrations. In LSE this is set to a real queue name, including here so we
 # can use settings.SERVICE_QUEUE_NAME in async migrations in LSO
